+++ Arma3ObjectBuilder/ui/import_export_mcfg.py
@@ -44,12 +44,14 @@ class A3OB_OP_import_mcfg(bpy.types.Operator, bpy_extras.io_utils.ImportHelper):
         count_skeletons = 0
         try:
             count_skeletons = import_mcfg.read_file(self, context)
+        except import_mcfg.rap.RAP_Error as ex:
+            utils.op_report(self, {'ERROR'}, "%s (check the system console)" % ex)
         except Exception as ex:
-            self.report({'ERROR'}, "%s (check the system console)" % str(ex))
+            utils.op_report(self, {'ERROR'}, "%s (check the system console)" % ex)
             traceback.print_exc()
         
         if count_skeletons > 0:
-            self.report({'INFO'}, "Successfully imported %d skeleton(s)" % count_skeletons)
+            utils.op_report(self, {'INFO'}, "Successfully imported %d skeleton(s)" % count_skeletons)
         
         return {'FINISHED'}
 
@@ -108,13 +110,17 @@ class A3OB_OP_export_mcfg(bpy.types.Operator, bpy_extras.io_utils.ImportHelper):
         pass
 
     def execute(self, context):
+        if not utils.OutputManager.can_access_path(self.filepath):
+            utils.op_report(self, {'ERROR'}, "Cannot write to target file (file likely in use by another blocking process)")
+            return {'FINISHED'}
+        
         scene_props = context.scene.a3ob_rigging
         skeleton = scene_props.skeletons[self.skeleton_index]
         output = utils.OutputManager(self.filepath, "w")
 
         validator = Validator(ProcessLoggerNull())
         if not validator.validate_skeleton(skeleton, False, True):
-            self.report({'ERROR'}, "Invalid skeleton definiton, run skeleton validation for more info")
+            utils.op_report(self, {'ERROR'}, "Invalid skeleton definiton, run skeleton validation for more info")
             return {'FINISHED'}
         
         with output as file:
@@ -122,7 +128,7 @@ class A3OB_OP_export_mcfg(bpy.types.Operator, bpy_extras.io_utils.ImportHelper):
                 export_mcfg.write_file(self, skeleton, file)
                 output.success = True
             except Exception as ex:
-                self.report({'ERROR'}, "%s (check the system console)" % str(ex))
+                utils.op_report(self, {'ERROR'}, "%s (check the system console)" % ex)
                 traceback.print_exc()
 
         return {'FINISHED'}