+++ Arma3ObjectBuilder/utilities/validator.py
@@ -7,11 +7,16 @@ import math
 import bmesh
 import bpy
 
+from ..io.data_p3d import P3D_LOD_Resolution as LOD
+
 
 class ValidatorResult():
     def __init__(self, is_valid = True, comment = ""):
         self.is_valid = is_valid
         self.comment = comment
+    
+    def __bool__(self):
+        return self.is_valid
     
     def set(self, is_valid = True, comment = ""):
         self.is_valid = is_valid
@@ -39,13 +44,13 @@ class ValidatorComponent():
 
         for item in strict:
             result = item()
-            if not result.is_valid:
+            if not result:
                 return False
         
         if warns_errs:
             for item in optional:
                 result = item()
-                if not result.is_valid:
+                if not result:
                     return False
         
         return True
@@ -57,13 +62,13 @@ class ValidatorComponent():
 
         for item in strict:
             result = item()
-            if not result.is_valid:
+            if not result:
                 self.logger.step("ERROR: %s" % result.comment)
                 is_valid = False
         
         for item in optional:
             result = item()
-            if not result.is_valid:
+            if not result:
                 self.logger.step("WARNING: %s" % result.comment)
                 is_valid &= not warns_errs
         
@@ -132,18 +137,11 @@ class ValidatorComponentLOD(ValidatorComponent):
             return False
         
         RE_NAME = re.compile(re_selection, re.IGNORECASE)
-        groups = [group.index for group in self.obj.vertex_groups if RE_NAME.match(group.name)]
+        for group in self.obj.vertex_groups:
+            if RE_NAME.match(group.name):
+                return True
         
-        if len(groups) == 0:
-            return False
-        
-        layer = self.bm.verts.layers.deform.verify()
-        for vert in self.bm.verts:
-            for group in groups:
-                if group in vert[layer] and vert[layer][group] == 1:
-                    return True
-        else:
-            return False
+        return False
     
     def only_ascii_vgroups(self):
         result = ValidatorResult()
@@ -213,7 +211,7 @@ class ValidatorLODGeometry(ValidatorComponentLOD):
 
     def is_triangulated(self):
         result =  super().is_triangulated()
-        if result.is_valid:
+        if not result:
             result.set(False, "mesh is not triangulated (convexity is not definite)")
         
         return result
@@ -371,6 +369,26 @@ class ValidatorLODShadow(ValidatorComponentLOD):
             self.no_materials
         )
         optional = info = tuple()
+
+        return strict, optional, info
+
+
+class ValidatorLODUnderground(ValidatorLODGeometry, ValidatorLODShadow):
+    """LOD - Underground (VBS)"""
+
+    def conditions(self):
+        strict = (
+            self.is_contiguous,
+            self.is_triangulated,
+            self.is_convex,
+            self.has_components,
+            self.is_sharp,
+            self.no_materials
+        )
+        optional = tuple()
+        info = (
+            self.farthest_point,
+        )
 
         return strict, optional, info
 
@@ -597,12 +615,44 @@ class Validator():
     
     def setup_lod_specific(self):
         self.components = {
-            **dict.fromkeys(("4", "26", "27", "28"), [ValidatorLODShadow]),
-            **dict.fromkeys(("9", "10", "13"), [ValidatorLODPointcloud]),
-            **dict.fromkeys(("7", "8", "14", "15", "16", "17", "19", "20", "21", "22", "23", "24"), [ValidatorLODGeometrySubtype]),
-            "6": [ValidatorLODGeometry],
-            "11": [ValidatorLODRoadway],
-            "12": [ValidatorLODPaths]
+            **dict.fromkeys(
+                (
+                    str(LOD.SHADOW),
+                    str(LOD.SHADOW_VIEW_CARGO),
+                    str(LOD.SHADOW_VIEW_PILOT),
+                    str(LOD.SHADOW_VIEW_GUNNER)
+                ), 
+                [ValidatorLODShadow]
+            ),
+            **dict.fromkeys(
+                (
+                    str(LOD.MEMORY),
+                    str(LOD.LANDCONTACT),
+                    str(LOD.HITPOINTS)
+                ),
+                [ValidatorLODPointcloud]
+            ),
+            **dict.fromkeys(
+                (
+                    str(LOD.GEOMETRY_BUOY),
+                    str(LOD.GEOMETRY_PHYSX),
+                    str(LOD.VIEW_GEOMETRY),
+                    str(LOD.FIRE_GEOMETRY),
+                    str(LOD.VIEW_CARGO_GEOMERTRY),
+                    str(LOD.VIEW_CARGO_FIRE_GEOMETRY),
+                    str(LOD.VIEW_COMMANDER_GEOMETRY),
+                    str(LOD.VIEW_COMMANDER_FIRE_GEOMETRY),
+                    str(LOD.VIEW_PILOT_GEOMETRY),
+                    str(LOD.VIEW_PILOT_FIRE_GEOMETRY),
+                    str(LOD.VIEW_GUNNER_GEOMETRY),
+                    str(LOD.VIEW_GUNNER_FIRE_GEOMETRY)
+                ),
+                [ValidatorLODGeometrySubtype]
+            ),
+            str(LOD.GEOMETRY): [ValidatorLODGeometry],
+            str(LOD.ROADWAY): [ValidatorLODRoadway],
+            str(LOD.PATHS): [ValidatorLODPaths],
+            str(LOD.UNDERGROUND): [ValidatorLODUnderground]
         }
 
     def validate_lod(self, obj, lod, lazy = False, warns_errs = True, relative_paths = False):
@@ -623,6 +673,7 @@ class Validator():
             is_valid &= item(obj, bm, self.logger, relative_paths).validate(lazy, warns_errs)
 
         bm.free()
+        self.logger.step("Validation %s" % ("PASSED" if is_valid else "FAILED"))
         self.logger.level_down()
         self.logger.step("Finished validation")
 
@@ -639,7 +690,8 @@ class Validator():
 
         for item in components:
             is_valid &= item(skeleton, self.logger).validate(lazy, False)
-
+        
+        self.logger.step("Validation %s" % ("PASSED" if is_valid else "FAILED"))
         self.logger.level_down()
         self.logger.step("Finished validation")